requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
twilio==8.10.0
pyrebase4==4.7.1
geopy==2.4.1
requests==2.31.0
pymongo==4.6.0